Turned BURT_can into a proper Arduino-compatible library
