+++ keywords.txt
@@ -1,0 +1,12 @@
+# Data types (KEYWORD1)
+BurtCan	KEYWORD1
+CanMessage KEYWORD1
+CanHandler KEYWORD1
+
+# Methods and functions (KEYWORD2)
+setup	KEYWORD2
+update	KEYWORD2
+send	KEYWORD2
+registerHandler	KEYWORD2
+
+# Constants (LITERAL1)
