+++ library.properties
@@ -1,0 +1,8 @@
+name=BURT CAN bus
+version=1.0.0
+author=Levi Lesches <levilesches@gmail.com>
+maintainer=Levi Lesches <levilesches@gmail.com>
+sentence=Handles sending and receiving messages using the CAN protocol for the Binghamton University Rover Team's subsystems.
+paragraph=Allows sketches to register handler functions that are triggered on certain message types, and allows for simple sending of messages. 
+category=Device Control
+url=https://github.com/BinghamtonRover/burt-can
